+++ docs/en/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## 6.0.3
 `To be released`
-+ 🆕 Add configuration `candle.tooltip.defaultValue`, `technicalIndicator.tooltip.defaultValue` and `separator.activeBackgroundColor`.
-+ 💄 Optimize drag and drop to adjust chart height display.
++ 🆕 Add configuration `candle.margin`, `candle.tooltip.defaultValue`, `technicalIndicator.margin`, `technicalIndicator.tooltip.defaultValue` and `separator.activeBackgroundColor`.
++ 💄 Optimize drag and drop to adjust chart display.
 + 🐞 Fix the problem of custom colors in `candle.tooltip.values`.